pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "bcrypt>=4.2.1",
    "numpy>=2.2.2",
    "pandas>=2.2.3",
    "plotly>=6.0.0",
    "psycopg2-binary>=2.9.10",
    "python-jose[cryptography]>=3.3.0",
    "sqlalchemy>=2.0.38",
    "streamlit>=1.42.0",
]
